requirements.txt
flask==3.0.3
APScheduler==3.10.4
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pandas==2.2.2
yfinance==0.2.52
requests==2.32.3
PyYAML==6.0.2
discord.py==2.3.2
websocket-client>=1.6.0
